add bazzite-gnome build (#9)

+++ build_files/build.sh
@@ -10,25 +10,26 @@ KERNEL_VERSION="$(rpm -q kernel --queryformat '%{VERSION}-%{RELEASE}.%{ARCH}')"
 
 ## Install tuxedo drivers
 # TODO: Remove this once https://github.com/ublue-os/akmods/pull/329 is merged
-dnf -y copr enable gladion136/tuxedo-drivers-kmod
-dnf -y install akmods \
-    "akmod-tuxedo-drivers-*.fc${RELEASE}.${ARCH}"
+# skip if we build for bazzite
+if [[ ! "${BASE_IMAGE}" =~ bazzite ]]; then
+    dnf -y copr enable gladion136/tuxedo-drivers-kmod
+    dnf -y install akmods \
+        "akmod-tuxedo-drivers-*.fc${RELEASE}.${ARCH}"
 
-akmods --force --kernels "${KERNEL_VERSION}" --kmod "tuxedo-drivers-kmod"
+    akmods --force --kernels "${KERNEL_VERSION}" --kmod "tuxedo-drivers-kmod"
 
-for file in /usr/lib/modules/${KERNEL_VERSION}/extra/tuxedo-drivers/*.ko.xz; do
-    modinfo "$file" > /dev/null \
-    || (find /var/cache/akmods/tuxedo-drivers/ -name \*.log -print -exec cat {} \; && exit 1)
-done
+    for file in /usr/lib/modules/${KERNEL_VERSION}/extra/tuxedo-drivers/*.ko.xz; do
+        modinfo "$file" > /dev/null \
+        || (find /var/cache/akmods/tuxedo-drivers/ -name \*.log -print -exec cat {} \; && exit 1)
+    done
 
-userdel akmods
-dnf -y remove akmods
-dnf -y copr disable gladion136/tuxedo-drivers-kmod
+    userdel akmods
+    dnf -y remove akmods
+    dnf -y copr disable gladion136/tuxedo-drivers-kmod
+fi
 
 ## Install cosmic related stuff
 dnf -y copr enable ryanabx/cosmic-epoch
 
 ## We just install cosmic session, because we run it alongside gnome for now
 dnf -y install cosmic-session
-
-dnf -y copr disable ryanabx/cosmic-epoch
